+++ src/main/kotlin/in/kkkev/jjidea/jj/cli/CliExecutor.kt
@@ -7,6 +7,7 @@ import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.util.Key
 import com.intellij.openapi.vcs.FilePath
 import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.util.containers.addAllIfNotNull
 import `in`.kkkev.jjidea.jj.*
 import `in`.kkkev.jjidea.vcs.pathRelativeTo
 import `in`.kkkev.jjidea.vcs.relativeTo
@@ -298,10 +299,21 @@ class CliExecutor(
     override fun gitClone(source: String, destination: String, colocate: Boolean) =
         execute(null, gitCloneArgs(source, destination, colocate), timeout = networkTimeout)
 
-    override fun configGet(key: String) = execute(null, listOf("config", "get", key))
+    override fun configGet(key: String) = execute(
+        root,
+        buildList { addAllIfNotNull("config", "get", key) }
+    )
 
-    override fun configSetUser(key: String, value: String) =
-        execute(null, listOf("config", "set", "--user", key, value))
+    override fun configList(key: String?, scope: CommandExecutor.ConfigScope?) = execute(
+        root,
+        buildList { addAllIfNotNull("config", "list", scope?.param, key) }
+    )
+
+    override fun configSetUser(scope: CommandExecutor.ConfigScope, key: String, value: String) =
+        execute(root, listOf("config", "set", scope.param, key, value))
+
+    override fun configUnset(scope: CommandExecutor.ConfigScope, key: String) =
+        execute(root, listOf("config", "unset", scope.param, key))
 
     /**
      * Clone a Git repository with streaming progress updates.